Tesla Slayer Exciter circuit LT Spice simulation
This works well on simulation for 10Vdc input 1.2kv to -1.2kv output is achieved. Diode is important to protect gate of power transistor. While simulating first drag 2 inductors from components section don't go to mirror or flip the components then very important is that k l1 l2 -1 need to be specified in spice directive. Also 0.1pF capacitor is needed to be grounded on the output side for coil to show the resonating effect.